+++ lib/src/source_visitor.dart
@@ -2100,7 +2100,7 @@ class SourceVisitor extends ThrowingAstVisitor
 
         if (node.elseStatement != null)
         {
-            if (_formatter.outerIfStatementElse)
+            if (_formatter.options.outerIfStatementElse)
             {
                 //! CHANGED(tekert) always add newline
                 newline();
@@ -3104,7 +3104,7 @@ class SourceVisitor extends ThrowingAstVisitor
             node.rightParenthesis);
 
         //! CHANGED(tekert): add new line on switch statements blocks.
-        if (_formatter.outerBracesOnBlockLike)
+        if (_formatter.options.outerBracesOnBlockLike)
         {
             writePrecedingCommentsAndNewlines(node.leftBracket);
             builder = builder.startBlock(space: false, indent: false);
@@ -3229,6 +3229,17 @@ class SourceVisitor extends ThrowingAstVisitor
     @override
     void visitSwitchStatement(SwitchStatement node)
     {
+        //! CHANGED(tekert): format empty body with no comments.
+        if (node.members.isEmptyBody(node.rightBracket))
+        {
+            // Don't allow splitting an empty switch expression.
+            _visitSwitchValue(node.switchKeyword, node.leftParenthesis, node.expression,
+                node.rightParenthesis);
+            token(node.leftBracket);
+            token(node.rightBracket);
+            return;
+        }
+
         _visitSwitchValue(node.switchKeyword, node.leftParenthesis, node.expression,
             node.rightParenthesis);
         //! CHANGED(tekert): add new argument node type
@@ -3363,7 +3374,7 @@ class SourceVisitor extends ThrowingAstVisitor
         space();
         visit(node.body);
 
-        if (_formatter.outerTryStatementClause)
+        if (_formatter.options.outerTryStatementClause)
         {
             //! CHANGED(tekert) newline before and between catch, on, finally, space removed.
             visitNodes(node.catchClauses, before: newline, between: newline);
@@ -4534,9 +4545,7 @@ class SourceVisitor extends ThrowingAstVisitor
             /* Treat EnumDeclaration  like literals for now. */
             (nodeType is! EnumDeclaration) &&
             /* SwitchPatternCase are aready open as of ^2.3.0 */
-            (nodeType.parent is! SwitchPatternCase) &&
-            /* NOTE: Dont Expand empty SwitchStatement for now 'switch(1) {}'. */
-            (nodeType is SwitchStatement ? nodeType.members.isNotEmpty : true))
+            (nodeType.parent is! SwitchPatternCase))
         {
             return true;
         }
@@ -4556,7 +4565,7 @@ class SourceVisitor extends ThrowingAstVisitor
         //! CHANGED(tekert) add new line on everything except some collection literals
         //!   and (Assertion In contructors and statements, ArgumentList with trailing comma).
         //! SwitchPatternCase Blocks by default uses a new line.
-        if (_formatter.outerBracesOnBlockLike &&
+        if (_formatter.options.outerBracesOnBlockLike &&
             (leftBracket.type == TokenType.OPEN_CURLY_BRACKET) &&
             _isBlockLike(nodeType))
         // EnumDeclaration is handled in [visitEnumDeclaration]